+++ substrate/frame/revive/src/tests/Playground.bin-runtime
@@ -1,0 +1,1 @@
+608060405234801561000f575f5ffd5b5060043610610034575f3560e01c806336ef737d14610038578063c6c2ea1714610056575b5f5ffd5b610040610086565b60405161004d91906100f2565b60405180910390f35b610070600480360381019061006b9190610139565b61008d565b60405161007d91906100f2565b60405180910390f35b5f43905090565b5f6001821161009e578190506100d5565b6100b36002836100ae9190610191565b61008d565b6100c86001846100c39190610191565b61008d565b6100d291906101c4565b90505b919050565b5f819050919050565b6100ec816100da565b82525050565b5f6020820190506101055f8301846100e3565b92915050565b5f5ffd5b610118816100da565b8114610122575f5ffd5b50565b5f813590506101338161010f565b92915050565b5f6020828403121561014e5761014d61010b565b5b5f61015b84828501610125565b91505092915050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52601160045260245ffd5b5f61019b826100da565b91506101a6836100da565b92508282039050818111156101be576101bd610164565b5b92915050565b5f6101ce826100da565b91506101d9836100da565b92508282019050808211156101f1576101f0610164565b5b9291505056fea26469706673582212204ac30a407dbf9ed259f7731524d17df5d6cad6bff1e4efdee4f6ffe34226f70e64736f6c634300081e0033
